backend/requirements-railway.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI/ML Libraries - Optional, will gracefully degrade if not available
openai==1.3.7

# Computer Vision - Headless for server
opencv-python-headless==4.10.0.84
Pillow==11.0.0
numpy==1.26.4

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
pydantic-settings==2.1.0

